Initial commit of actual program code
When this program is executed, it will connect to an IMAP server
for an account (which must first be configured; see the block of
constants at the top of green_shirt.py) and download all unread  messages
that have "[GREEN SHIRT]" in the subject. It will then attempt to
use Google Translate to convert the first 100 characters to speech
and speak it aloud. After processing all downloaded emails, it will
quit.