Don't split up words when chunking long messages
If a chunk (except the last one) ends in an alpha-numeric character,
we'll move the end leftward until we hit a non-alpha-numeric character.
This isn't the smartest behavior, since it's possible that we happened
to split the chunk originally at a word boundary, but it does stop
us from spliting up words.